tools/fill refactoring
