+++ package.json
@@ -81,6 +81,7 @@
     "webpack-node-externals": "^1.6.0"
   },
   "dependencies": {
-    "material-ui": "^0.20.0"
+    "material-ui": "^0.20.0",
+    "node-couchdb": "^1.3.0"
   }
 }
